Materials - changed pieces/eaches types from decimal to int

+++ Inventory/Inventory.Core/Entities/Material.cs
@@ -10,7 +10,7 @@ namespace Inventory.Core.Entities
         public virtual string PartNumber { get; set; }
         public virtual string Description { get; set; }
         public virtual MaterialType Type { get; set; }
-        public virtual decimal PiecesPerCase { get; set; }
-        public virtual decimal EachesPerPiece { get; set; }
+        public virtual int PiecesPerCase { get; set; }
+        public virtual int EachesPerPiece { get; set; }
     }
 }
